requirements.txt
pandas>=2.2.3
openpyxl==3.1.2
requests==2.31.0
lxml>=5.2.0
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
psycopg2-binary>=2.9.0
pytest>=7.4.0
pytest-cov>=4.1.0

